gcp-functions/validar-estado-tarjeta/requirements.txt
requests==2.31.0
google-cloud-logging
google-cloud-pubsub